(domain-graph)  Move domain graph to link-graph API module
