+++ code/functions/link-graph/api/java/nu/marginalia/domaingraph/DomainGraphBuilder.java
@@ -1,4 +1,4 @@
-package nu.marginalia.domainranking.data;
+package nu.marginalia.domaingraph;
 
 import it.unimi.dsi.fastutil.ints.IntArrayList;
