+++ code/functions/link-graph/api/java/nu/marginalia/domaingraph/GraphSource.java
@@ -1,4 +1,4 @@
-package nu.marginalia.domainranking.data;
+package nu.marginalia.domaingraph;
 
 import java.util.List;
