+++ code/index/java/nu/marginalia/index/searchset/SearchSetsService.java
@@ -10,9 +10,9 @@ import nu.marginalia.db.DomainTypes;
 import nu.marginalia.domainranking.PageRankDomainRanker;
 import nu.marginalia.domainranking.accumulator.RankingResultHashMapAccumulator;
 import nu.marginalia.domainranking.accumulator.RankingResultHashSetAccumulator;
-import nu.marginalia.domainranking.data.GraphSource;
-import nu.marginalia.domainranking.data.LinkGraphSource;
-import nu.marginalia.domainranking.data.SimilarityGraphSource;
+import nu.marginalia.domaingraph.GraphSource;
+import nu.marginalia.domaingraph.LinkGraphSource;
+import nu.marginalia.domaingraph.SimilarityGraphSource;
 import nu.marginalia.index.IndexFactory;
 import nu.marginalia.index.searchset.connectivity.ConnectivitySets;
 import nu.marginalia.index.searchset.connectivity.ConnectivityView;
