+++ code/index/test/nu/marginalia/domainranking/RankingAlgorithmsContainerTest.java
@@ -4,10 +4,10 @@ package nu.marginalia.domainranking;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import nu.marginalia.api.linkgraph.AggregateLinkGraphClient;
-import nu.marginalia.domainranking.data.DomainGraph;
-import nu.marginalia.domainranking.data.InvertedLinkGraphSource;
-import nu.marginalia.domainranking.data.LinkGraphSource;
-import nu.marginalia.domainranking.data.SimilarityGraphSource;
+import nu.marginalia.domaingraph.DomainGraph;
+import nu.marginalia.domaingraph.InvertedLinkGraphSource;
+import nu.marginalia.domaingraph.LinkGraphSource;
+import nu.marginalia.domaingraph.SimilarityGraphSource;
 import nu.marginalia.test.TestMigrationLoader;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.parallel.Execution;
