+++ code/index/test/nu/marginalia/domainranking/TestGraphSourceForInvertedLinkData.java
@@ -1,9 +1,9 @@
 package nu.marginalia.domainranking;
 
 import nu.marginalia.array.LongArrayFactory;
-import nu.marginalia.domainranking.data.DomainGraph;
-import nu.marginalia.domainranking.data.DomainGraphBuilder;
-import nu.marginalia.domainranking.data.GraphSource;
+import nu.marginalia.domaingraph.DomainGraph;
+import nu.marginalia.domaingraph.DomainGraphBuilder;
+import nu.marginalia.domaingraph.GraphSource;
 import org.apache.commons.lang3.StringUtils;
 
 import java.nio.file.Files;
